portrait-enhancer/requirements.txt
pillow
pyyaml
requests
# ADetailer 2CN Plus integration
numpy>=1.24.0
opencv-python>=4.8.0
pydantic>=2.0.0
onnxruntime-gpu>=1.15.0
insightface>=0.7.3
facenet-pytorch>=0.2.1
matplotlib>=3.7.0
